requirements.txt
fastapi==0.68.1
openai==0.27.0
python-dotenv==0.19.1
dateparser==1.1.8
psycopg2-binary==2.9.6
